+++ v2/frontend/src/pages/portfolio/hooks/usePortfolio.ts
@@ -1,5 +1,5 @@
 import { useQuery } from '@tanstack/react-query'
-import { fetchPortfolioSummary, fetchPortfolioHistory } from '../api/portfolio'
+import { fetchPortfolioSummary, fetchPortfolioHistory } from '../../../api/portfolio'
 import {
   presentPortfolioSummary,
   portfolioValueSeries,
@@ -7,7 +7,7 @@ import {
   portfolioDrawdown,
   type RawPortfolioSummary,
   type RawPortfolioHistoryRow,
-} from '../presenters/portfolioPresenter'
+} from '../../../presenters/portfolioPresenter'
 
 export function usePortfolioSummary() {
   return useQuery({
